Don't make assembly references jumpable
Unity callstacks include references to guids instead of files for
internal Unity code. Skip these so we don't get errors like:

 E303: Unable to open swap file for "<819de1aa368e45faa4f78e26c97c62b0>", recovery impossible

In theory, you could make vim edit the appropriate file instead of
skipping them, but you'd need UnityCsReference checked out and in a
specific location and the search for the right guid. If you did that,
you could either PR it here or do -= and += to support it locally. But
I'm not even use they're normal guids -- I don't see them in
UnityCsReference.